Eliminate stack to avoid memory leaks
This commit eliminates the stack of focused elements. Since React
operates on a tree, we get a stack for free as a by-product.

This prevents a case where the first active element is never removed
from the stack, causing a memory leak.

I've also done some small changes to the focus trap component to avoid
some extra work only required for testing.